[ 1816533 ] Rewrite the MutableBuffer class.
git-svn-id: https://svn.code.sf.net/p/oorexx/code-0/interpreter-3.x/trunk@1988 0b6cbdbe-3aab-466e-b73a-abd511dda0a2

+++ kernel/platform/unix/RexxQueues.cpp
@@ -61,7 +61,7 @@
 RexxMethod0(REXXOBJECT, rexx_query_queue)
 {
    REXXOBJECT queue_name;              /* current queue name                */
-   size_t count = 0;                   /* count of lines                    */
+   unsigned int count = 0;             /* count of lines                    */
    APIRET rc;                          /* queue query return code           */
 
                                        /* get the queue name                */
@@ -184,7 +184,7 @@ RexxMethod1(REXXOBJECT, rexx_create_queue,
 {
    char buf[name_parameter_length+1];  /* creation buffer                   */
    APIRET rc;                          /* creation return code              */
-   size_t dup_flag = 0;                /* duplicate name flag               */
+   unsigned int dup_flag = 0;          /* duplicate name flag               */
 
                                        /* create a queue                    */
    rc = RexxCreateQueue(buf, sizeof(buf), queue_name, &dup_flag);
